requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.0
coverage==7.6.9
decorator==5.1.1
distlib==0.3.9
Django==5.1.4
django-extensions==3.2.3
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
exceptiongroup==1.2.2
executing==2.1.0
extended-mypy-django-plugin==0.8.0
filelock==3.16.1
identify==2.6.3
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipython==8.30.0
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
mypy==1.13.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pygraphviz==1.14
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
wcwidth==0.2.13
